day00/ex04/volumes/app.volume/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
numpy